baitap-submit/ten_cua_ban/02-llm-api-params/requirements.txt
openai==1.58.1
python-dotenv==1.0.1
